+++ day11/bytecode.h
@@ -20,8 +20,8 @@ namespace AoC
         void RunBytecode(const std::vector<u8>& buffer);
         void Reset();
 
-        inline u32 GetRegisterA() const { return m_RegisterA; }
-        inline void SetRegisterA(u32 value) { m_RegisterA = value; }
+        inline u64 GetRegisterA() const { return m_RegisterA; }
+        inline void SetRegisterA(u64 value) { m_RegisterA = value; }
 
     private:
         void Execute(Bytecode operation, const std::vector<u8>& buffer, u32& instructionPointer);
@@ -31,7 +31,7 @@ namespace AoC
         void Execute_Add(const std::vector<u8>& buffer, u32& instructionPointer);
         void Execute_Mul(const std::vector<u8>& buffer, u32& instructionPointer);
 
-        std::vector<u32> m_Stack;
-        u32 m_RegisterA;
+        std::vector<u64> m_Stack;
+        u64 m_RegisterA;
     };
 }
